MAINT move log_loss and hinge_loss to the classification metrics
