+++ sklearn/metrics/__init__.py
@@ -6,11 +6,9 @@ and pairwise metrics and distance computations.
 from .ranking import auc
 from .ranking import average_precision_score
 from .ranking import label_ranking_average_precision_score
-from .ranking import log_loss
 from .ranking import precision_recall_curve
 from .ranking import roc_auc_score
 from .ranking import roc_curve
-from .ranking import hinge_loss
 
 from .classification import accuracy_score
 from .classification import classification_report
@@ -18,7 +16,9 @@ from .classification import confusion_matrix
 from .classification import f1_score
 from .classification import fbeta_score
 from .classification import hamming_loss
+from .classification import hinge_loss
 from .classification import jaccard_similarity_score
+from .classification import log_loss
 from .classification import matthews_corrcoef
 from .classification import precision_recall_fscore_support
 from .classification import precision_score
